rt-app: use per instance timer
when we instantiate a thread which contains timer events, the same
timers will be used for all instances and the instance will be woken up
sequentially.

Create a special value named "unique" for the "ref" key of timer. During
the creation of the event, the timer ref value will be replaced by a unique
value so each instance will use its own timer instead of sharing the same timer

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>