Do not sleep if we have run longer than expected
Add a check at the end of each job to see if we need to sleep
before the next one. Currently this is useless, as clock_nanosleep
already implements such behaviour. We anyway add it now, as it
could be useful in the future if, for example, we want to keep track
of how many times we ran more than expected.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Signed-off-by: Pi-Cheng Chen <pi-cheng.chen@linaro.org>