rt-app: fix calibration issue on mt8173
Calibration sequence fails to get the right ns per loop value when the
calibration is done on the A57 cluster of the mt8173evb.
A new calibration has been added that adds a sleep period between each each
calibration loop. This idle phase enable rt-app to get the to get the right ns
per loop value for such kind of HW.
The calibration sequence finally uses the lowest value of the calibration
methods which match with the highest compute capacity

Reported-by: Koan-Sin Tan <freedom.tan@linaro.org>
Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Reviewed-by: Amit Kucheria <amit.kucheria@linaro.org>