rt-app: reset timer that overruns
If the next wake up date of the timer is already in the past when we execute
the event, we resets the timer.t_next to the current time.
Otherwise, if the delay has been too large, we might never recover it and the
timer will always be in the past.