rt-app: add log buffer in memory
With the new mechanism that is used to describe a scenario, it becomes more
complex to estimate how many logs will be generated during the execution of a
use case. As a result, the record of the logs in temporary buffer had been
disable and the logs were saved directly in files. The potential side effect
of such mecanism is to block threads on io access, which disturbs the use case
behavior.

A new parameter is added to define the saving policy of the logs. You can
now disable the logs, save them directly in a file like previously,
define the size of a temporary buffer but with the risk to lost some logs if
the buffer overflow. You can also ask rt-app to evaluate how many logs will be
generated and allocate the temporary buffer accordingly. This last mode is not
yet implemented and it will be part of a dedicated patch.

Full details of the new parameter is available in the update of the
documentation

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>