workgen: Add IO-bounded and memory-bounded events
Add 2 new kind of event for running a memory or a io bounded load.
"mem" name for a load is memory bounded, and "iorun" name for a load is
io bounded. The default file to be written to create the load is
/dev/null and the device/file could be specified with "io_device" key in
"global" section. The size of per-thread memory buffer to create
IO/memory load could also be specified with "mem_buffer_size" in the
same section.

Signed-off-by: Pi-Cheng Chen <pi-cheng.chen@linaro.org>
[ replace Text by String in tutorial.txt]
Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>