rt-app: fix deadlock
If a sigterm is raised once we have installed our own handler but threads
are not already created, we deadlock in the shutdown function.

So, we don't wait anymore for thread until all threads has been created
which is not a big issue because we don't really care of gathering all
logs and traces before exiting rt-app in this case as the use case has
not started yet.

Reported-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>