rt-app: conditionally compile libdl
sched_setattr syscall has been added in v3.14 with deadline scheduler. This
new syscall is not always implemented in the lib and libdl is used in this
case.

We add a test on the presence of sched_setattr in the system environment and
compile libdl iff sched_setattr has not been already defined by another library

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>