Fix segfault if no resource is defined
The malloc() function returns non-initialized memory. If no locking is defined
for a task in the json config, then nblocages is not initialized, thus causing
a segfault. Fix this situation by initialize the memory to zero.

For a similar reason, set the sche_nice field of the sched_attr on the stack.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>