rt-parse: avoid repetition in obj_is_event()
obj_is_event() is a boolean function that returns true if the name is an
event.  The two users of the function use it as a boolean function.  The
function has a lot of unnecessary copy&paste and duplication.  Avoid
repeating ourselves and just loop through all the events in an array.

Signed-off-by: Javi Merino <javi.merino@arm.com>