Add taskgroup support for SCHED_OTHER tasks
The implementation is based on libcgroup. It can be optionally
configured by specifying --with-cgroup.

Building instruction and an example are given in doc/tutorial.txt.

Tested on x86_64 and arm64.

Known issue: The build shows some warnings that some of the functions
in libcgroup.a (e.g. cgroup_change_cgroup_flags()) are calling
functions (e.g. getgrgid()) which require at run-time the same
version of shared libraries from the glibc version used for linking.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>