Add Barrier Sync Primitive
I wanted a way to provide a sync point for two threads where it didn't
matter which one got there first - the first thread to reach the sync
point would wait, and the last would just signal and continue.

Using a bare lock/signal/wait/unlock on both sides results in the second
thread being suspended briefly immediately after it signals which is
not desirable when trying to mimic existing thread activation patterns.

If we don't use bare signal/wait then it doesn't need to be restricted
to just 2 threads, you can sync an arbitrary number (>1) of threads at
a given point in a use case by including barrier events in them.

This effect could be achieved with a pthread_barrier, but there is no
officially sanctioned way to cancel a barrier sync event once it has
started and no way to tell if any threads are waiting without adding
such a mechanism externally, so we use the pattern here instead.

Each barrier point has a mutex, a shared counter and a condition
variable.

When parsing the usecase, the shared counter is incremented for
each reference.

At barrier sync, the event locks the event-specific mutex and checks
to see if it is the last task. If it is not the last task, then it
waits on the event-specific condition variable. If it is the last
task, then it sends a broadcast to all other users of the condition
variable. The shared counter is used to determine when the last
task reaches the sync point.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>