rt-app: fix loadwait perf calculation
loadwait() returns perf. However, when exec is bigger than 1s it is
splitted in burts of execution to not overflow load_count variable. In
this case we return as perf the remaining exec (or 0 if exec is a
multiple of 1s!).

Fix it by precalculating, and then returning, the perf number.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>