parse_config: fix loop parsing when no phases are present
When no phases are present we parse the loop value (if present) from
parse_thread_phase_data(). At thread object level we are only interested
to know if a loop is specified or not (in which case we want to loop
forever).

Encode this information in data->loop by checking if loop property is
present.

Suggested-by: Vincent Guittot <vincent.guittot@linaro.com>
Signed-off-by: Juri Lelli <juri.lelli@arm.com>