rt-app: add per phase set priority event
Add capability to change task's priority when starting a new phase. The
behavior is similar to events. a change only happen when value are defined
in the phase otherwise the scheduling parameters stay unchanged.

All 4 scheduling classes are supported:
-Other
-Fifo
-RR
-deadline

We can change scheduling class, priority or deadline parameters.
The file tutorial.txt has been updated to describe this new capability and
the common properties of threads and phases have also been reordered a bit.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>