Parse tasks with 0 instance value
To introduce a fork event, we want to allow tasks with 0 instance values
to be referenced as well as tasks that are created and run at start
time.

To allow this we ensure we parse the thread data for 0 instance tasks so
that in next commit fork events can find it and re-use it.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>