fork: add a limit on allowable number of forks
To protect against infinite loop of forks, hard limit how many forks
each thread can do. Once the limit is reached we exit on error.

The limit is set arbitrarily. It was found that usually ulimit limits
the number of files that can be opened by a thread and likely we would
have hit this limit when creating log files for new threads before we
hit this fork limit.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>