Fix a segfault on SIGINT and refactor shutdown()
When SIGINT is called while pthread_join() is already called in main()
we could endup calling join twice on the same thread which is undefined
behaviour and leads to segfaults within pthread_join() that are not
meaningful.

This was caught while testing fork bombs and doing CTRL+c to force early
termination. Weirdly the segfault was happening inside pthread_join()
but when joining thread at index 57 but the interrupted pthread_join() at
main was joining thread at index 2. It might be interrupting join to do
another join is causing some issues especially the pthread_join() wihtin
shutdown() would have done pthread_join() on thread index 2 before it
has reached index 57.

+-- main()
   +-- pthread_join()         <---+
   SIGINT                         |
   +-- shutdown()                 +------ undefined behaviour
      +-- pthread_join()      <---|

Introduce a mutex to protect against this and refactor shutdown() to
make the code easier to follow.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>