taskgroups: Data structures and bricks for config parser
A taskgroup_data_t is allocated on demand from an array of
taskgroup_data_t's with 32 entries during configuration parsing.
In case that a same taskgroup is configured multiple times the same
taskgroup_data_t is referenced. The number of supported taskgroups can
be increased by changing max_nbr_tgs at compile time.

taskgroup_data_t::offset is used to know which part of a taskgroup
directory is owned by the taskgroup_data_t. E.g. in case
taskgroup_data_t::name is /tg1/tg11/tg111 and the taskgroup
/tg1/tg11 exists already in the system, taskgroup_data_t::offset is
/tg1/tg11.
The offset functionality makes sure that the taskgroup_data_t array does
not have to be sorted for taskgroup removal.

A taskgroup will be handled like CPU affinity, scheduling policy or
task priority. I.e. in case it is specified, it is valid for the whole
duration of the thread or phase. It is not handled like an event (e.g.
timer or barrier).

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>