taskgroups: Thread/phase taskgroup config check
To avoid a scenario in which a thread phase runs in different taskgroups
make sure that a taskgroup is not specified on thread level if phases
are specified.

A configuration which would provoke such behaviour otherwise is the
following:

    "tasks" : {
        "thread0" : {
            "taskgroup" : "/tg1"
            "phases" : {
                "phase0" : {
                },
                "phase1" : {
                    "taskgroup" : "/tg1/tg11"
                }
            }
        }
    }

During the first loop "phase0" runs in "/tg1" whereas in all other loops
it runs in "/tg1/tg11".

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>