ftrace: Add support for ftrace categories and make them easy to parse
Currently trace events are almost a clone of the log statements which makes them
not so easy to parse by using generic regular expressions.
Moreover, we can only enable all the ftrace events together despite we have
they correspond to different tasks or workload stages.

Add the concept of ftrace categories and update the global tracing configuration
option to make it possible to define which trace events should be injected
in the trace buffer.

Update events definition to use a "unique word" followed by "key=value" pairs.
Do that by removing redundant information like the task id which is already
encoded in the task name.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>

---
NOTE: documentation and JSON updates follow in a dedicated patches