rt-app/config: Keep track of event names
For each event we generated a custom description starting from its name
and other task specific attributes. However, the unique name is never stored.

Add the required support to keep track, for each event of a task, of its
(unique) name.
This can be useful, for example, to decorate ftrace events.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>