rtapp/config: Fix "ftrace" for backward support
The "ftrace" JSON option has always been considered optional and, by default, no
ftrace events was genererated if that option was not specified.

With the addition of the ftrace category support in:

   commit a2d7d8c571eb ("ftrace: Add support for ftrace categories and make them easy to parse")

and specifically with code to grant backward compatibility for the now
deprecated bool type, the JSON parser complains if a value for "ftrace" is not
specified, thus failing to be completely backward compatibile.

Fix that by adding a check on "ftrace" missing and allowing the code to smoothly
fall-back to the default behavior.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>