rt-app: Fix unchanged priority
It is possible to declare in an agenda a phase with no policy or priority
change, but with another dl parameter or uclamp change e.g.

  "phases": {
              "0" {
                   "policy": "SCHED_OTHER"
                   ... },
              "1" {
                    "util_min": 512
                   ... },
   },

Phase "0" will init the policy and the NICE value to 0. In phase "1", the
priority (i.e NICE) isn't changed and therefore, prio = -1. However, as
util_min is set, the parsed sched_data is valid. During the phase "1"
runtime, rt-app will then set NICE to -1 before setting "util_min" to 512.

Bringing THREAD_PRIORITY_UNCHANGED prio value, to let rt-app be aware of
unchanged prio and avoid setting -1 (which is valid or not, depending on
the sched policy) when the agenda doesn't modify the "priority" key.

Meanwhile, renaming set_thread_priority to set_thread_param, as the real
purpose of the function isn't just priority. Refactor as well the entire
function to improve readability and ease the THREAD_PRIORITY_UNCHANGED
handling.

Signed-off-by: Vincent Donnefort <vincent.donnefort@arm.com>