parse_config: Fix wrong priority info for SCHED_DEADLINE task
thread_body() issues a log_notice() printing task policy and priority:

  rt-app dl.json
  [rt-app] <notice> [0] Starting with SCHED_DEADLINE policy with priority 10

This is due to the fact that parse_sched_data() uses
DEFAULT_THREAD_PRIORITY (10) to initialize prior_def for a
SCHED_DEADLINE task. This value is only valid for a SCHED_FIFO or
SCHED_RR task.

Use prio_def=0 for a SCHED_DEADLINE task and change from the if/else to
a switch statement since it is clearer.
Use designated structure initialization for tmp_data.policy to save an
else path.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>