config: parse_task_event_data(): Bail on malformed config
When the JSON configuration is not matching the expected schema, the
current behavior is to print an error log and carry on with a 0-duration
run event. This is quite harmful as execution therefore continues
(potentially unnoticed), with a workload that is not at all what the
user wanted.

This is especially important when:
* many phases are involved. The error can be lost in the flow
* on automated setups: if no one is there to watch the log, critical
  errors like malformed input must be translated to a negative exit
  status.

Since there is no valid reason to provide an ill-formed (and therefore
meaningless) input, exit when this case is detected, which will make the
problem obvious to the user.