rt-app: core scheduling implementation rt-app_types: added core_scheduling_families_count and core_scheduling_family options rt-app_parse_config: added core_scheduling_families_count and core_scheduling_family parser from config file
The basic idea behind core scheduling is that two thread with different core scheduling cookie cannot run simultaneously on different cpu that are physically on the same core.
So when two or more threads have the same core_scheduling_family it will result having the same cookie, otherwise they will have different core scheduling cookie.

It has been implemented for the moment by maintaining a list of the first thread pid for every family. When a new thread with same family will be launched,
it will copy the cookie from the saved pid in the list.

We intend to change this implementation in future commits to allow core_scheduling_family changes in phases.