Update build documentation
numactl has switched to cmake and therefore the build instructions
had to be updated. Other minor adaptions were done, too.

numactl when used by file path needs to be in the numactl directory,
so that was updated. Furthermore the git clone was removed to be
in line with the other github link.

Signed-off-by: Christian Loehle <christian.loehle@arm.com>