rt-app: Mask cleanup signals on all threads but main
Ensure that the cleanup signals are only handled over the main thread.
This will be needed in order to safely take the fork lock when canceling
all threads on shutdown. The fork lock is only taken by tasks not the
main thread, thus it will become safe to take also from a signal handler
if that one is only run over main.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>