rt-app: Cancel all threads on shutdown, rather than kicking resources
Trying to get all threads out of their blocking state via signaling
condition variables while not even holding related locks is just racy.
At the same time, it is actually simpler to just cancel all threads.

On the one side, the hard cancellation will prevent some "closing words"
of the threads in the logs. On the other, it gives race-free termination
of threads, preventing that we may get stuck while try to join never
finishing threads. The latter should be of higher value.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>